v5_new_email/complaint_container/requirements.txt
-i https://artifacts-read.gkp.jpmchase.net/artifactory/api/pypi/pypi/simple
datasets>=2.10.1
numpy==1.24.4
scikit-learn==1.2.2
scipy==1.9.3
nltk==3.7
imbalanced-learn==0.10.1
lightgbm==3.3.5
catboost==1.1.1
xgboost==1.7.1
seaborn==0.12.2
Jinja2==3.1.2
ipywidgets
tqdm==4.64.0
#fastapi==0.100.0
#uvicorn==0.23.1
fastapi~=0.74.0
uvicorn==0.17.0
requests~=2.25.0
PyYAML==5.3.1
joblib==1.3.1
pandas==1.5.0
pytest==7.0.1
pytest-cov==3.0.0
jsonschema==3.2.0
python-multipart